make business specific api clients, delete clients for each request type, make business specific tests without steps classes
