--- src/main/java/client/DeleteApiClient.java
@@ -1,10 +1,0 @@
-package client;
-import model.User;
-import io.restassured.response.Response;
-
-public class DeleteApiClient extends BaseHttpClient{
-
-    public Response deleteUser(String accessToken) {
-        return doDeleteRequestWithAccessToken(super.baseUrl + "/api/auth/user",accessToken);
-    }
-}
